divided code into smaller components and added utils folder. cleaned up unused code
