+++ emotion_tracker/src/components/TabletEmotionButtons.js
@@ -1,55 +1,21 @@
 import EmotionStats from "./EmotionStats";
 import React, { useState, useEffect } from "react";
 import '../css/EmotionButtons.css';
-import emotionData from "../data/emotionData";
-
-
-const getButtonClassName = (label) => {
-  // Generate a unique class name based on the button's label
-  return `${label}`;
-};
+import {timerStart, timerTick} from "./TimerFunctions";
+import EmotionButtons from "./EmotionButtons";
 
 
 
-function TabletEmotionButton({ updateStats, setViewCondition, viewCondition }) {
-  const [statsData, setStatsData] = useState();
-  const [statsTodayData, setStatsTodayData] = useState();
+function TabletEmotionButton() {
   const [update, setUpdate] = useState(false);
-  const [timerText, setTimerText] = useState("")
   const [buttonActive, setButtonActive] = useState(true);
   const [time, setTime] = useState(0);
   const timerTimeMs = 1000;
-  const [startAnimation, setStartAnimation] = useState(false);
   const [clicked, setClicked] = useState(0)
-
-  // TIMER
-  const timerStart = (e) => {
-    console.log("timerStart ");
-    e.preventDefault();
-      setButtonActive(false);
-      const nyt = Date.now()
-      localStorage.setItem('timer', nyt);
-  };
-
-  const timerTick = () => {
-    // console.log("timerTick ~ ");
-    if (localStorage.getItem('timer')){
-      let res = Date.now() - localStorage.getItem('timer');
-      setTime((timerTimeMs-res) > 0 ? timerTimeMs-res : 0);
-      // console.log(res);
-
-      if (res > timerTimeMs){
-        setButtonActive(true);
-        setClicked(0)
-      } else {
-        setButtonActive(false);
-      }
-    }
-  };
 
   useEffect(() => {
     let timer = setInterval(() => {
-      timerTick();
+      timerTick(setTime, timerTimeMs, setButtonActive, setClicked);
     }, 1000);
     return () => clearInterval(timer);
   }, []);
@@ -72,52 +38,29 @@ function TabletEmotionButton({ updateStats, setViewCondition, viewCondition }) {
       }
       // handle success response here
       setUpdate(!update)
-      // setStatsData()
-      // setStatsTodayData("kku")
     } catch (error) {
+      console.log(`${error} error addEmotion`)
       // handle error here
     }
   };
 
 
   const buttonClicked = async (id, e) => {
-    // button clicked
-
     console.log("button clicked " + id);
     addEmotion(id);
     setClicked(id);
-    timerStart(e);
-
+    timerStart(e, setButtonActive);
       }
 
-  // const [disable, setDisable] = useState(false)
 
   return (
     <div className="content">
-      <div className="emotion-buttons">
-        {emotionData.map((button) => (
-          <button
-            style={{animation: startAnimation ? "fadeIn 3s, forwards" : "none"}}
-            key={button.label}
-            className={clicked !== button.id && !buttonActive ? getButtonClassName(button.label + "-disabled") : getButtonClassName(button.label)}
-            id = {clicked === button.id ? getButtonClassName(button.label + "-clicked") : getButtonClassName(button.label)}
-            disabled={!buttonActive}
-            onClick={(e) => buttonClicked(button.id, e)}
-          >
-            <div className="EmotionButton-button-label">
-              <span className="material-symbols-outlined">{button.icon}</span>
-              {button.label}
-            </div>
-          </button>
-        ))}
-            </div>
+    <EmotionButtons buttonActive={buttonActive} clicked={clicked} buttonClicked={buttonClicked}></EmotionButtons>
             <div style={{visibility: buttonActive ? "hidden" : "visible"}}>
-          <p class="infoText">
+          <p className="infoText">
           </p>
           </div>
       <EmotionStats
-        // statsData={statsData}
-        // statsTodayData={statsTodayData}
         update={update}
       />
     </div>
